reworked rss publisher to keep status of a configurable number recent builds.
Converted over to using the ruby rss library.  There are better libraries
that could be used, but this one comes with ruby so it doesn't add new external
dependencies.

A configuration option 'keep' specifies how many total builds to retain in the feed.
The default is 1 to match the previous behavior.

The rss library really insists on building a valid feed, so I had to add
description and link elements that weren't present in the previous code.
The channel_link should be a url set in your config, or it will default to
using a file url pointed at the rss file being written.
The description is the string that used to be the title, and the new
title is now shorter.

The generated rss file will look slightly different to a human, but it
should be valid and it passes the Cerberus rss_publisher_test.rb
In some cases you might need to make adjustments in whatever scripts you
may have that parse the rss after it's generated.